add pattern reference in TECommonControls. update references to TEColorParameter
